[cocoa] Support _hostOverride for redirects and WebSockets https://bugs.webkit.org/show_bug.cgi?id=278927 rdar://135021328
Reviewed by NOBODY (OOPS!).

Currently the _hostOverride SPI is only set on the data task when it is
created. This is problematic for some redirects when the port number changes,
and this means we completely miss WebSockets. In this patch, I'm moving most of
the logic into the shared super class of those two types of network requests,
NetworkTask. I'm also adjusting the override value when we handle a redirect.

No new tests.

* Source/WebKit/NetworkProcess/cocoa/NetworkDataTaskCocoa.mm:
(WebKit::NetworkDataTaskCocoa::NetworkDataTaskCocoa):
* Source/WebKit/NetworkProcess/cocoa/NetworkTaskCocoa.h:
* Source/WebKit/NetworkProcess/cocoa/NetworkTaskCocoa.mm:
(WebKit::NetworkTaskCocoa::willPerformHTTPRedirection):
(WebKit::NetworkTaskCocoa::setHostOverrideForTesting):
* Source/WebKit/NetworkProcess/cocoa/WebSocketTaskCocoa.mm:
(WebKit::WebSocketTask::WebSocketTask):