[css-view-transitions-2] no-view-transition-with-cross-origin-redirect.sub.html fails. https://bugs.webkit.org/show_bug.cgi?id=278837 <rdar://134908686>
Reviewed by Alex Christensen.

Cross-origin redirects should prevent the view transition from being allowed

This changes the test a bit, since 'hosts[][www]}' was resolving to
'www.localhost'. I'm not sure if that's a WebKit bug, or a test harness issue,
but using hosts[alt] should be a valid version of the test (where the initial
request is cross origin, but the post-redirect response is not).

It also moves the failure annotation for pageswap-long-delay.html into the
TestExpectations file, so that a URL to the spec issue can be added.

* LayoutTests/TestExpectations:
* LayoutTests/imported/w3c/web-platform-tests/css/css-view-transitions/navigation/no-view-transition-with-cross-origin-redirect.sub.html:
* LayoutTests/imported/w3c/web-platform-tests/css/css-view-transitions/navigation/pageswap-long-delay-expected.txt:
* Source/WebCore/loader/DocumentLoader.cpp:
(WebCore::DocumentLoader::navigationCanTriggerCrossDocumentViewTransition):

Canonical link: https://commits.webkit.org/283004@main