[Mac] run-minibrowser won't accept url as an argument https://bugs.webkit.org/show_bug.cgi?id=278955
Reviewed by Sam Weinig.

On Mac's MiniBrowser, it won't accept url as an launch argument so that
run-minibrowser's behavior is different from other platform.

Change implementation of AppDelegate to convert the file path argument to
url if possible. Existing `--url` is kept as is for compatibility.

* Tools/MiniBrowser/mac/AppDelegate.m:
(-[BrowserAppDelegate application:openFile:]):

Canonical link: https://commits.webkit.org/283063@main