window.open with noopener should start using a new BrowsingContextGroup when site isolation is enabled https://bugs.webkit.org/show_bug.cgi?id=278971 rdar://135085188
Reviewed by Charlie Wolfe.

When noopener is used, that means we don't wnat to be able to post a message to it.  If we
don't have an opener relationship but we are in the same BrowsingContextGroup, the process
management gets confused and starts putting things in the wrong process.

This is the first step towards getting fast/dom/Window/a-rel-noopener.html to work with --site-isolation.
It no longer asserts.  Next, I need to make it not time out.

I think I also need a followup PR to make it so there isn't a WebPage/Page created in the
opener process at all if it's a cross-site open with noopener just to return null from
window.open.

* Source/WebKit/UIProcess/WebPageProxy.cpp:
(WebKit::WebPageProxy::createNewPage):
* Tools/TestWebKitAPI/Tests/WebKitCocoa/SiteIsolation.mm:
(TestWebKitAPI::TEST(SiteIsolation, OpenWithNoopener)):

Canonical link: https://commits.webkit.org/283065@main