[ iOS MacOS ] http/tests/navigation/page-cache-shared-worker.html is a flaky timeout rdar://133720855 https://bugs.webkit.org/show_bug.cgi?id=277986
Reviewed by Chris Dumez.

LayoutTests/http/tests/navigation/resources/page-cache-helper-for-sharedworker.html was creating sw2 and sending it a message in a 50 ms async loop.
This is not needed as we only need to set the state once.
This may also add up to 100 events to sw2 which might delay the processing of the getState message (the one that is timing out), especially on slow bots.
We update the test to only set the hungry state once.

We almost make the test more robust by rejecting the test promises after 5 seconds with a message and making sure to call history.back even if sending the message pong is taking more than one second.

* LayoutTests/http/tests/navigation/page-cache-shared-worker.html:
* LayoutTests/http/tests/navigation/resources/page-cache-helper-for-sharedworker.html:
* LayoutTests/platform/ios/TestExpectations:
* LayoutTests/platform/ipad/TestExpectations:
* LayoutTests/platform/mac/TestExpectations:

Canonical link: https://commits.webkit.org/283085@main