Allow layout tests to load model content on visionOS with model process https://bugs.webkit.org/show_bug.cgi?id=279082 rdar://124870608
Reviewed by Mike Wyrzykowski and Per Arne Vollan.

Set up an entitlements file for the model process for simulator that have
the same two entitlements we added for the model process on device.

Make sure ModelElementEnabled is true for the model tests.
Make sure ModelProcessEnabled is true For the model tests that require
the model process to be able to process model content.

Fix the flakiness of model-element-error-and-load-events.html.

Also add simple layout tests for bounding box and entity transform.

* LayoutTests/model-element/model-element-bounding-box-expected.txt: Added.
* LayoutTests/model-element/model-element-bounding-box.html: Added.
* LayoutTests/model-element/model-element-camera.html:
* LayoutTests/model-element/model-element-enter-fullscreen-crash.html:
* LayoutTests/model-element/model-element-entity-transform-expected.txt: Added.
* LayoutTests/model-element/model-element-entity-transform.html: Added.
* LayoutTests/model-element/model-element-error-and-load-events.html:
The first test case sets the model's source to a valid URL and then removes the source.
There's a chance that the model source finished loading, queued the load event
to be fired, and then its source was removed. By the time the load event is
actually fired, model.complete is already false. Update the assert for the load
event to account for that case, and update model_load_test to take in an array
of possible intermediate events that can happen before the expected event is
fired. This way we can pass in the load event in the possibleIntermediateEvents
array for this test case.
* LayoutTests/model-element/model-element-inline-preview-deletion-upon-source-change.html:
* LayoutTests/model-element/model-element-interactive-dragging.html:
* LayoutTests/model-element/model-element-interactive.html:
* LayoutTests/model-element/model-element-ready.html:
* LayoutTests/model-element/model-element-source.html:
* LayoutTests/model-element/model-element.html:
* LayoutTests/model-element/resources/2x2x2-at-100x50x200.usdz: Added.
* LayoutTests/model-element/resources/2x2x2-center.usdz: Added.
* LayoutTests/model-element/resources/2x2x2-positive.usdz: Added.
* LayoutTests/model-element/resources/model-utils.js:
(const.assert_points_are_equal):
(const.assert_3d_matrix_is_identity):
(const.assert_3d_matrix_equals):
(const.assert_3d_matrix_not_equals):
* Source/WebKit/Configurations/ModelService.xcconfig:
* Source/WebKit/Resources/ios/ModelService-embedded-simulator.entitlements: Added.
Very similar to the XPCService-embedded-simulator.entitlements with the added SurfBoard
entitlements needed by the model process.

Canonical link: https://commits.webkit.org/283159@main