Video is sometimes mispositioned in element fullscreen window https://bugs.webkit.org/show_bug.cgi?id=279102 rdar://135224530
Reviewed by Wenson Hsieh.

In webkit.org/b/278842 we found that the client may sometimes synchronously modify WKWebView's
_obscuredInsets when the web view is inserted in the element fullscreen window. 282915@main
attempted to fix this but it was incomplete because, in addition to _obscuredInsets, the client may
also modify the scroll view's contentInset and contentOffset. The result is similar to the issue
seen in bug 278842: the video is mispositioned in the fullscreen window (and sometimes can be
scrolled).

This change follows up 282915@main with a more complete fix. Instead of resetting just
_obscuredInsets and _unobscuredSafeAreaInsets after inserting the web view, we now just defer
resetting all WKWebView and WKScrollView properties until after -insertSubview:atIndex: is called.
Additionally, we reset all properties again right before changing fullscreenState to
WKFullscreenStateEnteringFullscreen since the client may modify properties while we wait for a
presentation update. At the point where fullscreenState changes to
WKFullscreenStateEnteringFullscreen we expect the client to avoid modifying WKWebView and
WKScrollView properties for the duration of the element fullscreen presentation.

* Source/WebKit/UIProcess/ios/fullscreen/WKFullScreenWindowControllerIOS.mm:
(-[WKFullScreenWindowController _enterFullScreen:windowScene:]):

Canonical link: https://commits.webkit.org/283160@main