Expose focusWebViewFromServiceWorker on iOS https://bugs.webkit.org/show_bug.cgi?id=279104 rdar://135247073
Reviewed by Per Arne Vollan.

`_focusWebViewFromServiceWorker` is marked as being available on iOS, but doesn't actually work
since it is part of a `PLATFORM(MAC)` block. Fix this by moving this out of the `PLATFORM(MAC)`
block.

For backwards compatibility, if the delegate doesn't implement this method, then continue to do
nothing on iOS. On Mac, we continue to provide a default implementation that tries to focus the web
view's window.

* Source/WebKit/UIProcess/API/Cocoa/WKUIDelegatePrivate.h:
* Source/WebKit/UIProcess/Cocoa/UIDelegate.h:
* Source/WebKit/UIProcess/Cocoa/UIDelegate.mm:
(WebKit::UIDelegate::setDelegate):
(WebKit::UIDelegate::UIClient::focusFromServiceWorker):

Canonical link: https://commits.webkit.org/283162@main