[SVG] zoom applied twice to font size in foreignObject https://bugs.webkit.org/show_bug.cgi?id=279041
Reviewed by NOBODY (OOPS!).

WebKit handles zoom factor (zoom property and page zoom) differently
for SVG. WebKit multiplies computed CSS length values by a zoom factor
for HTML elements, but for SVG elements. WebKit scales SVG
differently. Thus, font-size property is zoomed for HTML elements, but
unzoomed for a SVG elements.

If a compound HTML document included a SVG document that included
another HTML document, and font-size property was specified to a HTML
element of the outer HTML document, and a zoom factor was applied, the
computed font-size was zoomed. If it was inherited to elements of the
inner HTML document. WebKit scaled SVG. Eventually, the font size of
the inner HTML was zoomed twice.

font-size property should be resolved at root elements of included
documents.

* LayoutTests/svg/zoom/page/zoom-foreignObject-2-expected.html: Added.
* LayoutTests/svg/zoom/page/zoom-foreignObject-2.html: Added.
* Source/WebCore/css/svg.css:
(*|*:not(*) > svg, foreignObject > *|*:not(*)):